+++ src/modules/checkout/config/keys/localStorageKey.ts
@@ -1,0 +1,1 @@
+export const PRODUCT_CART_INFO_KEY = 'product_cart_info_key';