+++ src/modules/core/providers/W3blockSDKGeneralConfigProvider/W3blockSDKGeneralConfigProvider.tsx
@@ -1,6 +1,7 @@
 import { ReactNode, useMemo } from 'react';
 
 import { PixwayUISdkLocale } from '../../context';
+import { EnvironmentContext } from '../../context/EnvironmentContext';
 import { W3blockUISDKGereralConfigContext } from '../../context/W3blockUISDKGeneralConfigContext';
 import { LocaleProvider } from '../LocaleProvider';
 import { W3blockApiProvider } from '../W3blockApiProvider';
@@ -10,10 +11,13 @@ interface Props {
   api: {
     idUrl: string;
     keyUrl: string;
+    commerceUrl: string;
   };
   locale: PixwayUISdkLocale;
   companyId: string;
   logoUrl: string;
+  isProduction: boolean;
+  appBaseUrl: string;
 }
 
 export const W3blockUISDKGeneralConfigProvider = ({
@@ -22,17 +26,32 @@ export const W3blockUISDKGeneralConfigProvider = ({
   locale,
   companyId,
   logoUrl,
+  isProduction,
+  appBaseUrl,
 }: Props) => {
-  const value = useMemo(() => ({ companyId, logoUrl }), [logoUrl, companyId]);
+  const companyValue = useMemo(
+    () => ({ companyId, logoUrl, appBaseUrl }),
+    [logoUrl, companyId, appBaseUrl]
+  );
+
+  const environmentValue = useMemo(
+    () => ({
+      isProduction,
+    }),
+    [isProduction]
+  );
 
   return (
-    <W3blockUISDKGereralConfigContext.Provider value={value}>
-      <W3blockApiProvider
-        w3blockIdAPIUrl={api.idUrl}
-        w3blockKeyAPIUrl={api.keyUrl}
-      >
-        <LocaleProvider locale={locale}>{children}</LocaleProvider>
-      </W3blockApiProvider>
+    <W3blockUISDKGereralConfigContext.Provider value={companyValue}>
+      <EnvironmentContext.Provider value={environmentValue}>
+        <W3blockApiProvider
+          w3blockIdAPIUrl={api.idUrl}
+          w3blockKeyAPIUrl={api.keyUrl}
+          w3blockCommerceAPIUrl={api.commerceUrl}
+        >
+          <LocaleProvider locale={locale}>{children}</LocaleProvider>
+        </W3blockApiProvider>
+      </EnvironmentContext.Provider>
     </W3blockUISDKGereralConfigContext.Provider>
   );
 };