+++ src/modules/shared/interface/Currency/index.ts
@@ -1,0 +1,9 @@
+import { CurrencyEnum } from '../../enums/Currency';
+
+export interface Currency {
+  id: string;
+  createdAt: string;
+  updatedAt: string;
+  name: CurrencyEnum;
+  crypto: boolean;
+}