+++ src/modules/shared/interface/index.ts
@@ -1,0 +1,5 @@
+export * from './Currency';
+
+export * from './Price';
+
+export * from './Product';